requirements.txt
# ─── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# ─── Redis ────────────────────────────────────────────────────────────────────
redis[asyncio]==5.2.1

# ─── Auth ─────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0

# ─── LangChain / LangGraph ────────────────────────────────────────────────────
langchain==0.3.12
langchain-openai==0.2.12
langchain-core>=0.3.25,<0.4.0
langgraph>=0.2.55,<0.3.0

# ─── Milvus ───────────────────────────────────────────────────────────────────
pymilvus[model]>=2.5.0,<2.6.0

# ─── HTTP Client ──────────────────────────────────────────────────────────────
httpx==0.28.1

# ─── Document Processing ──────────────────────────────────────────────────────
pdfplumber>=0.11.0
python-docx==1.1.2
chardet==5.2.0

# ─── Settings ─────────────────────────────────────────────────────────────────
pydantic-settings==2.6.1
email-validator>=2.1.0
pydantic==2.10.3
python-dotenv==1.0.1

# ─── Utilities ────────────────────────────────────────────────────────────────
aiofiles==24.1.0
numpy==1.26.4
scipy==1.14.1
jieba==0.42.1
rank-bm25==0.2.2

# ─── Evaluation ───────────────────────────────────────────────────────────────
ragas==0.2.6
datasets==3.2.0
